// packages/nestjs/src/integrations/vendored/enums/AttributeNames.ts
/*
 * Copyright The OpenTelemetry Authors
 * SPDX-License-Identifier: Apache-2.0
 *
 * NOTICE from the Sentry authors:
 * - Vendored from: https://github.com/open-telemetry/opentelemetry-js-contrib/tree/15ef7506553f631ea4181391e0c5725a56f0d082/packages/instrumentation-nestjs-core
 * - Upstream version: @opentelemetry/instrumentation-nestjs-core@0.64.0
 */

export enum AttributeNames {
  VERSION = 'nestjs.version',
  TYPE = 'nestjs.type',
  MODULE = 'nestjs.module',
  CONTROLLER = 'nestjs.controller',
  CALLBACK = 'nestjs.callback',
}